split module Index.Compact
